+++ packages/eslint-config-eslint/index.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const baseConfigs = require("./base");
-const { esmConfigs } = require("./node");
+const { esmConfigs } = require("./nodejs");
 
 module.exports = [
     ...baseConfigs,
